Changed staging buffer file sizes to unsigned long long int

+++ include/Prefetcher.h
@@ -19,7 +19,7 @@ public:
                int seed);
     ~Prefetcher();
     char* get_staging_buffer();
-    int get_next_file_end();
+    unsigned long long int get_next_file_end();
     int get_dataset_length();
 
 private:
@@ -36,7 +36,7 @@ private:
     std::vector<std::map<std::string, std::string>> config_pf_backend_options;
     std::vector<std::map<int, int>> config_bandwidths;
     std::map<int, int> config_pfs_bandwidth;
-    std::deque<int> file_ends;
+    std::deque<unsigned long long int> file_ends;
     std::vector<std::vector<std::thread>> threads;
     int node_id = 0;
